+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/typescript/Typescript.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2020 DiffPlug
+ * Copyright 2016-2023 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@ import com.diffplug.spotless.maven.FormatterFactory;
 /**
  * A {@link FormatterFactory} implementation that corresponds to {@code <typescript>...</typescript>} configuration element.
  * <p>
- * It defines a formatter for typescript source files.
+ * It defines formatters for typescript source files.
  */
 public class Typescript extends FormatterFactory {
 	@Override
@@ -38,5 +38,9 @@ public class Typescript extends FormatterFactory {
 
 	public void addTsfmt(Tsfmt tsfmt) {
 		addStepFactory(tsfmt);
+	}
+
+	public void addEslint(EslintTs eslint) {
+		addStepFactory(eslint);
 	}
 }